feat: mapper

+++ src/main/java/br/com/fiap/study_apir/controller/ProdutoController.java
@@ -3,6 +3,7 @@ package br.com.fiap.study_apir.controller;
 import java.util.List;
 import java.util.Optional;
 
+import org.apache.catalina.connector.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
@@ -19,55 +20,67 @@ import org.springframework.web.bind.annotation.RequestBody;
 
 import br.com.fiap.study_apir.dto.ProdutoCreateRequest;
 import br.com.fiap.study_apir.dto.ProdutoMapper;
+import br.com.fiap.study_apir.dto.ProdutoResponse;
+import br.com.fiap.study_apir.dto.ProdutoUpdateRequest;
 import br.com.fiap.study_apir.model.Produto;
 import br.com.fiap.study_apir.repository.ProdutoRepository;
 import br.com.fiap.study_apir.repository.RepositoryProdutoMockup;
 
 import br.com.fiap.study_apir.service.ProdutoService;
+import jakarta.validation.Valid;
 
 // anotação para informar que é controller
 @RestController
 @RequestMapping("api/${api.version}/produtos")
 public class ProdutoController {
-   
+
     @Autowired
     private ProdutoService service;
 
     // nova injeção
     @Autowired
     private ProdutoMapper produtoMapper;
-   
+
     // criar método que responda as aplicações - CRUD
 
     // método POST
     @PostMapping
-    public ResponseEntity<Produto> create(@RequestBody ProdutoCreateRequest dtoRequest) {
-        return ResponseEntity.status(HttpStatus.CREATED).body(service.createOrUpdate(produtoMapper.toModel(dtoRequest))); 
+    public ResponseEntity<ProdutoResponse> create(@Valid @RequestBody ProdutoCreateRequest dtoRequest) {
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body(produtoMapper.toDto(service.createOrUpdate(produtoMapper.toModel(dtoRequest))));
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Produto> findById(@PathVariable Long id) {
+    public ResponseEntity<ProdutoResponse> findById(@PathVariable Long id) {
         return service
                 .findById(id)
-                .map(ResponseEntity::ok) 
-                .orElse(ResponseEntity.notFound().build()); 
+                // precisamos de um map adicional - vamos usar o mapper que vai converter o
+                // produto para um DTO
+                .map(produto -> produtoMapper.toDto(produto))
+                .map(ResponseEntity::ok)
+                .orElse(ResponseEntity.notFound().build());
     }
 
     // find all
     @GetMapping
-    public ResponseEntity<List<Produto>> findAll() {
-        return ResponseEntity.ok(service.findAll());
+    public ResponseEntity<List<ProdutoResponse>> findAll() {
+        return ResponseEntity.ok(service.findAll()
+                .stream()
+                // aqui fazemos a conversão - cria uma nova lista em memória (de DTOs)
+                .map(produto -> produtoMapper.toDto(produto))
+                .toList() // colocamos pois o map não é final - é intermediário
+        );
     }
 
-    // método PUT 
+    // método PUT
     @PutMapping("/{id}")
-    public ResponseEntity<Produto> update(@PathVariable Long id, @RequestBody Produto produto) { 
-       Optional<Produto> optProduto = service.findById(id);
+    public ResponseEntity<ProdutoResponse> update(@PathVariable Long id,
+            @RequestBody @Valid ProdutoUpdateRequest dtoRequest) {
 
-       if(optProduto.isPresent()){
-        produto.setId(id);
-        Produto produtoAlterado = service.createOrUpdate(produto); 
-        return ResponseEntity.ok(produtoAlterado);
+        if (service.findById(id).isPresent()) {
+            Produto produto = produtoMapper.toModel(id, dtoRequest);
+            produto.setId(id);
+            return ResponseEntity.ok(produtoMapper.toDto(service.createOrUpdate(produto)));
         } else {
             return ResponseEntity.notFound().build();
         }
@@ -75,8 +88,12 @@ public class ProdutoController {
 
     // método DELETE
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteById(@PathVariable Long id) { 
-        service.deleteById(id);
-        return ResponseEntity.noContent().build(); 
+    public ResponseEntity<Void> deleteById(@PathVariable Long id) {
+        if (service.findById(id).isPresent()) {
+            service.deleteById(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
